+++ espnet_onnx/export/asr/models/joint_network.py
@@ -4,10 +4,10 @@ import numpy as np
 import torch
 import torch.nn as nn
 
-from .abs_model import AbsModel
+from espnet_onnx.utils.abs_model import AbsExportModel
 
 
-class JointNetwork(nn.Module, AbsModel):
+class JointNetwork(nn.Module, AbsExportModel):
     def __init__(self, model, search_type):
         super().__init__()
         self.model = model
